move tests to `TestFileUtils` and setup mocking context
