+++ src/test/java/org/terasology/launcher/util/TestFileUtils.java
@@ -19,6 +19,10 @@ package org.terasology.launcher.util;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
+import org.junit.runner.RunWith;
+import org.powermock.api.mockito.PowerMockito;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.PowerMockRunner;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -33,7 +37,11 @@ import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
+@RunWith(PowerMockRunner.class)
+@PrepareForTest(FileUtils.class)
 public class TestFileUtils {
 
     private static final String FILE_NAME = "File";
@@ -42,6 +50,39 @@ public class TestFileUtils {
 
     @Rule
     public TemporaryFolder tempFolder = new TemporaryFolder();
+
+    @Test(expected = IOException.class)
+    public void testCannotCreateDirectory() throws IOException {
+        PowerMockito.mockStatic(Files.class);
+
+        final Path directory = mock(Path.class);
+        when(Files.exists(directory)).thenReturn(false);
+        when(Files.createDirectories(directory)).thenThrow(new IOException("Failed to create directories"));
+
+        FileUtils.ensureWritableDir(directory);
+    }
+
+    @Test(expected = IOException.class)
+    public void testNotDirectory() throws IOException {
+        PowerMockito.mockStatic(Files.class);
+
+        final Path directory = mock(Path.class);
+        when(Files.exists(directory)).thenReturn(true);
+        when(Files.isDirectory(directory)).thenReturn(false);
+
+        FileUtils.ensureWritableDir(directory);
+    }
+
+    @Test(expected = IOException.class)
+    public void testNoPerms() throws IOException {
+        PowerMockito.mockStatic(Files.class);
+
+        final Path directory = mock(Path.class);
+        when(Files.isReadable(directory)).thenReturn(false);
+        when(Files.isWritable(directory)).thenReturn(false);
+
+        FileUtils.ensureWritableDir(directory);
+    }
 
     @Test
     public void testDeleteFile() throws IOException {
